requirement.txt
psycopg2
python-multipart
python-jose
passlib[bcrypt]
python-dotenv
pydantic-settings
python-json-logger
bcrypt==3.2.0
fastapi-pagination
pytz
sqlmodel
fastapi
uvicorn
bcrypt
pydantic[email]